Add a list of tags by wallet. Wallet controller refactoring.
