+++ app/src/Controller/Wallets/ListController.php
@@ -1,0 +1,73 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Controller\Wallets;
+
+use App\Repository\WalletRepository;
+use App\Request\Wallet\SortSetRequest;
+use App\Service\Sort\SortService;
+use App\Service\Sort\SortType;
+use App\Service\UserOptionsService;
+use App\Service\UserService;
+use App\View\WalletsView;
+use Psr\Http\Message\ResponseInterface;
+use Spiral\Auth\AuthScope;
+use Spiral\Http\ResponseWrapper;
+use Spiral\Router\Annotation\Route;
+
+final class ListController extends Controller
+{
+    public function __construct(
+        AuthScope $auth,
+        private ResponseWrapper $response,
+        private WalletRepository $walletRepository,
+        private WalletsView $walletsView,
+        private UserOptionsService $userOptionsService,
+        private SortService $sortService,
+        private UserService $userService,
+    ) {
+        parent::__construct($auth);
+    }
+
+    #[Route(route: '/wallets', name: 'wallet.list', methods: 'GET', group: 'auth')]
+    public function list(): ResponseInterface
+    {
+        return $this->walletsView->json($this->walletRepository->findAllByUserPK((int) $this->user->id));
+    }
+
+    #[Route(route: '/wallets/unarchived', name: 'wallet.list.unarchived', methods: 'GET', group: 'auth', priority: 0)]
+    public function listUnArchived(): ResponseInterface
+    {
+        return $this->walletsView->json(
+            $this->walletRepository->findAllByUserPKByArchived((int) $this->user->id, false),
+            $this->userOptionsService->getSort($this->user, SortType::Wallets),
+        );
+    }
+
+    #[Route(route: '/wallets/unarchived/sort', name: 'wallet.sort.unarchived.set', methods: 'POST', group: 'auth', priority: 0)]
+    public function sortUnArchived(SortSetRequest $request): ResponseInterface
+    {
+        if (! $request->isValid()) {
+            return $this->response->json([
+                'errors' => $request->getErrors(),
+            ], 422);
+        }
+
+        try {
+            $this->sortService->set($this->user, SortType::Wallets, $request->getSort());
+            $this->userService->store($this->user);
+        } catch (\Throwable) {
+        }
+
+        return $this->response->create(200);
+    }
+
+    #[Route(route: '/wallets/archived', name: 'wallet.list.archived', methods: 'GET', group: 'auth', priority: 0)]
+    public function listArchived(): ResponseInterface
+    {
+        return $this->walletsView->json(
+            $this->walletRepository->findAllByUserPKByArchived((int) $this->user->id, true)
+        );
+    }
+}
