+++ app/src/Controller/Wallets/WalletsController.php
@@ -4,21 +4,13 @@ declare(strict_types=1);
 
 namespace App\Controller\Wallets;
 
-use App\Database\Charge;
 use App\Database\Currency;
 use App\Database\Wallet;
-use App\Repository\ChargeRepository;
 use App\Repository\CurrencyRepository;
 use App\Repository\WalletRepository;
 use App\Request\Wallet\CreateRequest;
-use App\Request\Wallet\SortSetRequest;
 use App\Request\Wallet\UpdateRequest;
-use App\Service\Sort\SortService;
-use App\Service\Sort\SortType;
-use App\Service\UserOptionsService;
-use App\Service\UserService;
 use App\Service\WalletService;
-use App\View\WalletsView;
 use App\View\WalletView;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
@@ -34,122 +26,13 @@ final class WalletsController extends Controller
         private LoggerInterface $logger,
         private WalletRepository $walletRepository,
         private WalletService $walletService,
-        private WalletsView $walletsView,
         private WalletView $walletView,
         private CurrencyRepository $currencyRepository,
-        private ChargeRepository $chargeRepository,
-        private UserOptionsService $userOptionsService,
-        private SortService $sortService,
-        private UserService $userService,
     ) {
         parent::__construct($auth);
     }
 
-    #[Route(
-        route: '/wallets',
-        name: 'wallet.list',
-        methods: 'GET',
-        group: 'auth',
-    )]
-    public function list(): ResponseInterface
-    {
-        return $this->walletsView->json($this->walletRepository->findAllByUserPK((int) $this->user->id));
-    }
-
-    #[Route(
-        route: '/wallets/unarchived',
-        name: 'wallet.list.unarchived',
-        methods: 'GET',
-        group: 'auth',
-    )]
-    public function listUnArchived(): ResponseInterface
-    {
-        return $this->walletsView->json(
-            $this->walletRepository->findAllByUserPKByArchived((int) $this->user->id, false),
-            $this->userOptionsService->getSort($this->user, SortType::Wallets),
-        );
-    }
-
-    #[Route(
-        route: '/wallets/unarchived/sort',
-        name: 'wallet.sort.unarchived.set',
-        methods: 'POST',
-        group: 'auth',
-    )]
-    public function sortUnArchived(SortSetRequest $request): ResponseInterface
-    {
-        if (! $request->isValid()) {
-            return $this->response->json([
-                'errors' => $request->getErrors(),
-            ], 422);
-        }
-
-        try {
-            $this->sortService->set($this->user, SortType::Wallets, $request->getSort());
-            $this->userService->store($this->user);
-        } catch (\Throwable) {
-        }
-
-        return $this->response->create(200);
-    }
-
-    #[Route(
-        route: '/wallets/archived',
-        name: 'wallet.list.archived',
-        methods: 'GET',
-        group: 'auth',
-    )]
-    public function listArchived(): ResponseInterface
-    {
-        return $this->walletsView->json($this->walletRepository->findAllByUserPKByArchived((int) $this->user->id, true));
-    }
-
-    #[Route(
-        route: '/wallets/<id>',
-        name: 'wallet.index',
-        methods: 'GET',
-        group: 'auth',
-    )]
-    public function index(int $id): ResponseInterface
-    {
-        $wallet = $this->walletRepository->findByPKByUserPK($id, (int) $this->user->id);
-
-        if (! $wallet instanceof Wallet) {
-            return $this->response->create(404);
-        }
-
-        return $this->walletView->json($wallet);
-    }
-
-    #[Route(
-        route: '/wallets/<id>/total',
-        name: 'wallet.index.total',
-        methods: 'GET',
-        group: 'auth',
-    )]
-    public function indexTotal(int $id): ResponseInterface
-    {
-        $wallet = $this->walletRepository->findByPKByUserPK($id, (int) $this->user->id);
-
-        if (! $wallet instanceof Wallet) {
-            return $this->response->create(404);
-        }
-
-        return $this->response->json([
-            'data' => [
-                'totalAmount' => $wallet->totalAmount,
-                'totalIncomeAmount' => $this->chargeRepository->totalByWalletPK($id, Charge::TYPE_INCOME),
-                'totalExpenseAmount' => $this->chargeRepository->totalByWalletPK($id, Charge::TYPE_EXPENSE),
-            ],
-        ]);
-    }
-
-    #[Route(
-        route: '/wallets',
-        name: 'wallet.create',
-        methods: 'POST',
-        group: 'auth',
-    )]
+    #[Route(route: '/wallets', name: 'wallet.create', methods: 'POST', group: 'auth')]
     public function create(CreateRequest $request): ResponseInterface
     {
         if (! $request->isValid()) {
@@ -170,12 +53,7 @@ final class WalletsController extends Controller
         return $this->walletView->json($wallet);
     }
 
-    #[Route(
-        route: '/wallets/<id>',
-        name: 'wallet.update',
-        methods: 'PUT',
-        group: 'auth',
-    )]
+    #[Route(route: '/wallets/<id>', name: 'wallet.update', methods: 'PUT', group: 'auth')]
     public function update(int $id, UpdateRequest $request): ResponseInterface
     {
         $wallet = $this->walletRepository->findByPKByUserPK($id, (int) $this->user->id);
@@ -234,12 +112,7 @@ final class WalletsController extends Controller
         return $this->walletView->json($wallet);
     }
 
-    #[Route(
-        route: '/wallets/<id>',
-        name: 'wallet.delete',
-        methods: 'DELETE',
-        group: 'auth',
-    )]
+    #[Route(route: '/wallets/<id>', name: 'wallet.delete', methods: 'DELETE', group: 'auth')]
     public function delete(int $id): ResponseInterface
     {
         $wallet = $this->walletRepository->findByPKByUserPK($id, (int) $this->user->id);
